migrated to jetpack datastore for authentication. Minor refactoring
