+++ app/src/main/java/org/privacymatters/safespace/LogActivity.kt
@@ -24,7 +24,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
 import org.privacymatters.safespace.main.ui.SafeSpaceTheme
-import org.privacymatters.safespace.utils.LockTimer
+import org.privacymatters.safespace.auth.LockTimer
 import org.privacymatters.safespace.utils.Utils
 import java.io.BufferedReader
 import java.io.File
