+++ fastlane/metadata/android/en-US/changelogs/47.txt
@@ -1,2 +1,2 @@
-Updated dependencies
-backstage improvements
+fixed green camera glitch on some phones
+app improvements
